ui/src/common/components/__tests__/ExpandListToggle.test.tsx
import React from 'react';
import { shallow } from 'enzyme';

import ExpandListToggle from '../ExpandListToggle';
import SecondaryButton from '../SecondaryButton';

// @ts-expect-error ts-migrate(2582) FIXME: Cannot find name 'describe'. Do you need to instal... Remove this comment to see the full error message
describe('ExpandListToggle', () => {
  // @ts-expect-error ts-migrate(2582) FIXME: Cannot find name 'it'. Do you need to install type... Remove this comment to see the full error message
  it('renders toggle size > limit', () => {
    const wrapper = shallow(
      <ExpandListToggle
        size={10}
        limit={5}
        // @ts-expect-error ts-migrate(2708) FIXME: Cannot use namespace 'jest' as a value.
        onToggle={jest.fn()}
        expanded={false}
      />
    );
    // @ts-expect-error ts-migrate(2304) FIXME: Cannot find name 'expect'.
    expect(wrapper).toMatchSnapshot();
  });

  // @ts-expect-error ts-migrate(2582) FIXME: Cannot find name 'it'. Do you need to install type... Remove this comment to see the full error message
  it('does not render toggle size == limit', () => {
    const wrapper = shallow(
      <ExpandListToggle
        size={5}
        limit={5}
        // @ts-expect-error ts-migrate(2708) FIXME: Cannot use namespace 'jest' as a value.
        onToggle={jest.fn()}
        expanded={false}
      />
    );
    // @ts-expect-error ts-migrate(2304) FIXME: Cannot find name 'expect'.
    expect(wrapper).toMatchSnapshot();
  });

  // @ts-expect-error ts-migrate(2582) FIXME: Cannot find name 'it'. Do you need to install type... Remove this comment to see the full error message
  it('does not render toggle size < limit', () => {
    const wrapper = shallow(
      <ExpandListToggle
        size={3}
        limit={5}
        // @ts-expect-error ts-migrate(2708) FIXME: Cannot use namespace 'jest' as a value.
        onToggle={jest.fn()}
        expanded={false}
      />
    );
    // @ts-expect-error ts-migrate(2304) FIXME: Cannot find name 'expect'.
    expect(wrapper).toMatchSnapshot();
  });

  // @ts-expect-error ts-migrate(2582) FIXME: Cannot find name 'it'. Do you need to install type... Remove this comment to see the full error message
  it('renders toggle with expanded true', () => {
    const wrapper = shallow(
      // @ts-expect-error ts-migrate(2708) FIXME: Cannot use namespace 'jest' as a value.
      <ExpandListToggle size={10} limit={5} onToggle={jest.fn()} expanded />
    );
    // @ts-expect-error ts-migrate(2304) FIXME: Cannot find name 'expect'.
    expect(wrapper).toMatchSnapshot();
  });

  // @ts-expect-error ts-migrate(2582) FIXME: Cannot find name 'it'. Do you need to install type... Remove this comment to see the full error message
  it('calls onToggle when button is clicked', () => {
    // @ts-expect-error ts-migrate(2708) FIXME: Cannot use namespace 'jest' as a value.
    const onToggle = jest.fn();
    const wrapper = shallow(
      <ExpandListToggle size={10} limit={5} onToggle={onToggle} expanded />
    );
    wrapper.find(SecondaryButton).simulate('click');
    // @ts-expect-error ts-migrate(2304) FIXME: Cannot find name 'expect'.
    expect(onToggle).toHaveBeenCalledTimes(1);
  });
});
